# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Buddies API
  version: 1.0.0
  description: |
    Dual-sided B2B reverse-auction marketplace API (consumer restaurants + suppliers).
    Auth: `Authorization: Bearer <supabase_jwt>` or dev `Bearer dev:<role>:<userId>`.
servers:
  - url: http://localhost:8000/v1
...
paths:
...
  /orders/{id}/return-claims:
    post:
      tags: [consumer]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Return claim draft }
  /return-claims/{id}/submit:
    post:
      tags: [consumer]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Submitted / auto-approved }
  /return-claims/{id}/supplier-decision:
    post:
      tags: [supplier]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [accept, dispute] }
                resolutionType: { type: string, enum: [refund, replacement] }
                refundAmountPaise: { type: integer }
                pickupWindow: { type: string }
                notes: { type: string }
      responses:
        '200': { description: Decision recorded. Accept + pickupWindow → pickup_scheduled. }
  /return-claims/{id}/schedule-pickup:
    post:
      tags: [supplier]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pickupWindow]
              properties:
                pickupWindow: { type: string }
      responses:
        '200': { description: Status pickup_scheduled }
  /return-claims/{id}/confirm-pickup:
    post:
      tags: [supplier]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Status picked_up }
  /return-claims/{id}/refund:
    post:
      tags: [supplier]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountPaise]
              properties:
                amountPaise: { type: integer, description: Offline cash/UPI amount in paise }
                receiptRef: { type: string }
      responses:
        '200': { description: Offline refund recorded (status refunded) }
  /return-claims/{id}/replaced:
    post:
      tags: [supplier]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Replacement delivered (status replaced) }
  /return-claims/{id}/consumer-ack:
    post:
      tags: [consumer]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Closes refunded, replaced, or rejected claims }
...
  /returns/policy:
    get:
      tags: [shared]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Policy summary }
  /returns/windows:
    get:
      tags: [shared]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Category windows }
  /catalog/shelf-life-matrix:
    get:
      tags: [shared]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: RSL matrix }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...